Changed how serializers were registered
- Moved FLUID_STACK_SERIALIZER to CBCDataSerializers and statically register them in CreateBigCannons
- Removed registration from platform specific mod classes
