+++ common/src/main/java/rbasamoyai/createbigcannons/CreateBigCannons.java
@@ -16,6 +16,7 @@ import rbasamoyai.createbigcannons.index.CBCBlockEntities;
 import rbasamoyai.createbigcannons.index.CBCBlocks;
 import rbasamoyai.createbigcannons.index.CBCChecks;
 import rbasamoyai.createbigcannons.index.CBCContraptionTypes;
+import rbasamoyai.createbigcannons.index.CBCDataSerializers;
 import rbasamoyai.createbigcannons.index.CBCEntityTypes;
 import rbasamoyai.createbigcannons.index.CBCFluids;
 import rbasamoyai.createbigcannons.index.CBCItems;
@@ -54,10 +55,11 @@ public class CreateBigCannons {
 	static {
 		REGISTRATE.setTooltipModifierFactory(item -> new ItemDescription.Modifier(item, TooltipHelper.Palette.STANDARD_CREATE)
 			.andThen(TooltipModifier.mapNull(KineticStats.create(item))));
+
+		CBCDataSerializers.registerSerializers();
 	}
 
 	public static ResourceLocation resource(String path) {
 		return new ResourceLocation(MOD_ID, path);
 	}
-
 }
